Moved classes to their own package
