+++ even-more-fish-plugin/src/main/java/com/oheers/fish/database/model/CompetitionReport.java
@@ -1,4 +1,4 @@
-package com.oheers.fish.database;
+package com.oheers.fish.database.model;
 
 import org.jetbrains.annotations.NotNull;
