+++ even-more-fish-plugin/src/main/java/com/oheers/fish/database/model/Transaction.java
@@ -1,4 +1,4 @@
-package com.oheers.fish.database;
+package com.oheers.fish.database.model;
 
 import java.sql.Timestamp;
